Amend time zone and file handling
I am in UTC+1 and was getting no lines matched because the log timestamp was being assumed local and then converted to universal. It should be assumed universal instead.

I was also getting exceptions thrown due to issueLogFile not being closed before it was zipped. Fixed by keeping the lines in RAM and writing them in one go at the end.

Finally, we only need to lock the log for as long as it takes to grab a copy.