Fix for annoying out-of-game galnet
Get a journal timestamp and use that to determine when it is appropriate to update the galnet monitor. Since updates will only be occurring in-game (or shortly after exiting game) now, we will poll a little more frequently.

Add a short delay after the monitor is activated before it begins pulling in new articles (time for commanders to settle into the game a bit before the Galnet Monitor starts)
Update change log for galnet changes.