Amend parsing of promotion ranks
This regression was my fault. Casts in C# can fail when unboxing numbers if the boxed number is not of the same type, even if a cast from the boxed type to th4e casted type would work. Apparently this is a performance optimization from the old days.

Fixed with unit tests by using `Convert.ToInt32()` instead.