#129 Optimize friends list handling
Prefer List<T> methods such as FindIndex() to manual for loops.

Prefer replacing in situ to removing and re-adding an item. For large lists this can have a significant performance impact as every element after the found item will have to be moved up and back down again (despite its name List<T> is implemented as a contiguous array not a linked list).