Revise TTS distortion effect per #326.
Make the distortion effect a modifier that increases the level of voice processing, rather than a separate effect. Thus, the effects level set by the user becomes a floor and the actual effects level increases with damage to the ship if this option is enabled.

Adjust audio gain for losses that occur as the effects level increases.