#424 fix startup crash with EDSM responder
It was attempting to sync even though the EDSM service was ot configured.
We now only start the update thread if the configuration is valid.
This is now done in `Reload()` rather than `Start()` to cater for cases where the user has just input a valid configuration.
Additionally we abort the update thread if the responder is told to stop.