#602 Fix SAR event commodity property
There seems to be some inconsistency here between documentation and reality. The Cottle scripts expect `commodity` to be an object with properties `name` and `rare`. The docs and wiki say that `commodity is a string, being the localized name of the commodity.

I have restored the object behavior as it is more useful, but I'm uncertain what consequences this will have to Voice Attack, so I really need some VA users to test it please!