Add variety to "Ship targeted" script
When in a combat-rich environment it was getting very repetitive, and too verbose. It would backlog the speech cue in some situatations where I was switching targets a lot.

This edit loses the redundant "Information :" prefix and adds some variation including the target's name, while keeping it fairly brief. I've tried it on some bounty-hunting sessions and it feels a lot better.