Added new ship and module defs (incomplete)
Unfortunately I did not capture EDIDs during the beta, so a lot here is provisional. It will at least allow EDDI to name the ships and modules correctly.

FR, DE and ES translations taken from the spreadsheet. They are abbreviated for the HUD and should be expanded for TTS purposes.