Non-default scripts in the Speech Responder UI
Resolves #1464.
Revise the `Delete` button in the Speech Responder UI to either `Delete` or `Reset`, as appropriate.
Store the default script when we're loading scripts and automatically determine whether a script matches the default.
If the `Delete` or `Reset` button is enabled then the script is non-default.
Adds a confirmation dialog prior to resetting a script from the main window tab.