[ru] Give the TTS full numbers rather than fragments and words
This lets the TTs decline numbers correctly for quantity (2-4 vs 5+ etc).

We give short decimals genitive singular but it is moot as that format is overridden and never used.

For the larger numbers (> trillions) the TTs voice limitations mean that we express the numbers in thousands of trillions, etc.

If we don't, the voices fall back to reading the numbers as just a sequence of digits, and that is no good.