Speculative fix for Rollbar #21533 - Inara event error: No name provided.
This issue seems to occur when someone tries to use the InaraDetails() function without providing a lookup name and without having configured an API key. With a personal API key, this endpoint would normally default to returning the personal data of the commander associated with the API key when no lookup name is given.